# [pdata] [chore] Reuse OneOf structs on copy