--- a/pdata/internal/wrapper_value.go
+++ b/pdata/internal/wrapper_value.go
@@ -30,6 +30,46 @@ func NewValue(orig *otlpcommon.AnyValue) Value {
 	return Value{orig: orig}
 }
 
+func CopyOrigValue(dst, src *otlpcommon.AnyValue) {
+	switch ov := src.Value.(type) {
+	case *otlpcommon.AnyValue_KvlistValue:
+		kv, ok := dst.Value.(*otlpcommon.AnyValue_KvlistValue)
+		if !ok {
+			kv = &otlpcommon.AnyValue_KvlistValue{KvlistValue: &otlpcommon.KeyValueList{}}
+			dst.Value = kv
+		}
+		if ov.KvlistValue == nil {
+			kv.KvlistValue = nil
+			return
+		}
+		// Deep copy to dest.
+		CopyOrigMap(&kv.KvlistValue.Values, &ov.KvlistValue.Values)
+	case *otlpcommon.AnyValue_ArrayValue:
+		av, ok := dst.Value.(*otlpcommon.AnyValue_ArrayValue)
+		if !ok {
+			av = &otlpcommon.AnyValue_ArrayValue{ArrayValue: &otlpcommon.ArrayValue{}}
+			dst.Value = av
+		}
+		if ov.ArrayValue == nil {
+			av.ArrayValue = nil
+			return
+		}
+		// Deep copy to dest.
+		CopyOrigSlice(&av.ArrayValue.Values, &ov.ArrayValue.Values)
+	case *otlpcommon.AnyValue_BytesValue:
+		bv, ok := dst.Value.(*otlpcommon.AnyValue_BytesValue)
+		if !ok {
+			bv = &otlpcommon.AnyValue_BytesValue{}
+			dst.Value = bv
+		}
+		bv.BytesValue = make([]byte, len(ov.BytesValue))
+		copy(bv.BytesValue, ov.BytesValue)
+	default:
+		// Primitive immutable type, no need for deep copy.
+		dst.Value = ov
+	}
+}
+
 func FillTestValue(dest Value) {
 	dest.orig.Value = &otlpcommon.AnyValue_StringValue{StringValue: "v"}
 }