--- a/pdata/pcommon/generated_byteslice.go
+++ b/pdata/pcommon/generated_byteslice.go
@@ -40,12 +40,12 @@ func NewByteSlice() ByteSlice {
 
 // AsRaw returns a copy of the []byte slice.
 func (ms ByteSlice) AsRaw() []byte {
-	return copyByteSlice(nil, *ms.getOrig())
+	return internal.CopyOrigByteSlice(nil, *ms.getOrig())
 }
 
 // FromRaw copies raw []byte into the slice ByteSlice.
 func (ms ByteSlice) FromRaw(val []byte) {
-	*ms.getOrig() = copyByteSlice(*ms.getOrig(), val)
+	*ms.getOrig() = internal.CopyOrigByteSlice(*ms.getOrig(), val)
 }
 
 // Len returns length of the []byte slice value.
@@ -98,10 +98,5 @@ func (ms ByteSlice) MoveTo(dest ByteSlice) {
 
 // CopyTo copies all elements from the current slice overriding the destination.
 func (ms ByteSlice) CopyTo(dest ByteSlice) {
-	*dest.getOrig() = copyByteSlice(*dest.getOrig(), *ms.getOrig())
-}
-
-func copyByteSlice(dst, src []byte) []byte {
-	dst = dst[:0]
-	return append(dst, src...)
+	*dest.getOrig() = internal.CopyOrigByteSlice(*dest.getOrig(), *ms.getOrig())
 }