This optimization removes allocations in case if a one of pdata field is copied to a destination with the same type

It includes https://github.com/open-telemetry/opentelemetry-collector/pull/7213


Before (run on https://github.com/open-telemetry/opentelemetry-collector/pull/7213):
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/collector/pdata/pmetric
BenchmarkMetricsCopyTo
BenchmarkMetricsCopyTo-10    	     805	   1272214 ns/op	 2286312 B/op	   42359 allocs/op
BenchmarkMetricsCopyToReuseDest
BenchmarkMetricsCopyToReuseDest-10    	     998	   1192810 ns/op	 2252868 B/op	   42189 allocs/op
```

After:
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/collector/pdata/pmetric
BenchmarkMetricsCopyTo
BenchmarkMetricsCopyTo-10    	     934	   1259999 ns/op	 2286317 B/op	   42359 allocs/op
BenchmarkMetricsCopyToReuseDest
BenchmarkMetricsCopyToReuseDest-10    	    6627	    178169 ns/op	       0 B/op	       0 allocs/op
```